third_party/WebKit/LayoutTests/fast/animation/scroll-animations/animation-created-with-scroll-timeline.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test an Animation created with a ScrollTimeline doesn't crash</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<script>
test(t => {
  // We don't yet support ScrollTimeline, but creating an Animation object with
  // one should not crash.
  var timeline = new ScrollTimeline({ timeRange: 100, orientation: 'block' });
  assert_throws("NotSupportedError", function() { new Animation(null, timeline) });
}, 'An Animation created with a ScrollTimeline should not crash');
</script>
